Add NetworkScenarioTypesLibrary
